Organization and cmd setup
Reorganized some of the code into a different directory structure. This
still needs to be thought through a little more. But, so far it makes
more sense to allow everything for a year to be in one directory.

Adding cmd to see how much easier it will be to test the input. Input in
Advent of Code is generally used for both parts of these puzzles. So,
we'll see how it pans out.